fastapi_app/requirements.txt
# FastAPI and ASGI server
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0

# Authentication & Security
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6

# WebSocket support
websockets>=11.0
redis>=4.0.0
aioredis>=2.0.0

# Validation & Serialization
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Development & Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0.0
factory-boy>=3.3.0

# Utilities
python-dotenv>=1.0.0
email-validator>=2.0.0
